+++ src/app/page.tsx
@@ -9,51 +9,34 @@ import { Slider } from "@/components/ui/slider";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import { useTheme } from "next-themes";
 import { ModeToggle } from "@/components/mode-toggle";
+import { ReactSortable } from "react-sortablejs";
+import { availableIcons, IconItem } from "../lib/icons"// Import from the new file
+import Image from "next/image";
 
-
-// NEW FILE BEING DEVLOPED IN src\app\dev\page.tsx
-
-
-const availableIcons = [
-  "ableton", "activitypub", "actix", "adonis", "ae", "aiscript", "alpinejs", "anaconda", "androidstudio", "angular",
-  "ansible", "apollo", "apple", "appwrite", "arch", "arduino", "astro", "atom", "au", "autocad",
-  "aws", "azul", "azure", "babel", "bash", "bevy", "bitbucket", "blender", "bootstrap", "bsd",
-  "bun", "c", "cs", "cpp", "crystal", "cassandra", "clion", "clojure", "cloudflare", "cmake",
-  "codepen", "coffeescript", "css", "cypress", "d3", "dart", "debian", "deno", "devto", "discord",
-  "bots", "discordjs", "django", "docker", "dotnet", "dynamodb", "eclipse", "elasticsearch", "electron", "elixir",
-  "elysia", "emacs", "ember", "emotion", "express", "fastapi", "fediverse", "figma", "firebase", "flask",
-  "flutter", "forth", "fortran", "gamemakerstudio", "gatsby", "gcp", "git", "github", "githubactions", "gitlab",
-  "gmail", "gherkin", "go", "gradle", "godot", "grafana", "graphql", "gtk", "gulp", "haskell",
-  "haxe", "haxeflixel", "heroku", "hibernate", "html", "htmx", "idea", "ai", "instagram", "ipfs",
-  "java", "js", "jenkins", "jest", "jquery", "kafka", "kali", "kotlin", "ktor", "kubernetes",
-  "laravel", "latex", "less", "linkedin", "linux", "lit", "lua", "md", "mastodon", "materialui",
-  "matlab", "maven", "mint", "misskey", "mongodb", "mysql", "neovim", "nestjs", "netlify", "nextjs",
-  "nginx", "nim", "nix", "nodejs", "notion", "npm", "nuxtjs", "obsidian", "ocaml", "octave",
-  "opencv", "openshift", "openstack", "p5js", "perl", "ps", "php", "phpstorm", "pinia", "pkl",
-  "plan9", "planetscale", "pnpm", "postgres", "postman", "powershell", "pr", "prisma", "processing", "prometheus",
-  "pug", "pycharm", "py", "pytorch", "qt", "r", "rabbitmq", "rails", "raspberrypi", "react",
-  "reactivex", "redhat", "redis", "redux", "regex", "remix", "replit", "rider", "robloxstudio", "rocket",
-  "rollupjs", "ros", "ruby", "rust", "sass", "spring", "sqlite", "stackoverflow", "styledcomponents", "sublime",
-  "supabase", "scala", "sklearn", "selenium", "sentry", "sequelize", "sketchup", "solidity", "solidjs", "svelte",
-  "svg", "swift", "symfony", "tailwind", "tauri", "tensorflow", "terraform", "threejs", "twitter", "ts",
-  "ubuntu", "unity", "unreal", "v", "vala", "vercel", "vim", "visualstudio", "vite", "vitest",
-  "vscode", "vscodium", "vue", "vuetify", "wasm", "webflow", "webpack", "webstorm", "windicss", "windows",
-  "wordpress", "workers", "xd", "yarn", "yew", "zig"
-];
 
 export default function IconBuilder() {
+  //Handle site theme
   const { theme, resolvedTheme } = useTheme();
   const selectedTheme = (resolvedTheme || theme) === "dark" ? "light" : "dark";
-  const [search, setSearch] = useState("");
-  const [selectedIcons, setSelectedIcons] = useState<string[]>([]);
-  const [filteredIcons, setFilteredIcons] = useState(availableIcons);
   const [isDarkTheme, setIsDarkTheme] = useState(true);
+
+  //Icon Searching and Selecting
+  const [search, setSearch] = useState("");
+  const [selectedIcons, setSelectedIcons] = useState<IconItem[]>([]);
+  const [filteredIcons, setFilteredIcons] = useState<IconItem[]>(availableIcons);
+  
+  // Icon Preview Options
+  const [iconTheme, setIconTheme] = useState(true);
   const [isCentered, setIsCentered] = useState(false);
   const [iconsPerLine, setIconsPerLine] = useState(10);
+  const [isCopied, setIsCopied] = useState(false);
 
   useEffect(() => {
     setFilteredIcons(
-      availableIcons.filter((icon) => icon.toLowerCase().includes(search.toLowerCase()))
+      availableIcons.filter((icon) =>
+        icon.name.toLowerCase().includes(search.toLowerCase()) ||
+        icon.id.toLowerCase().includes(search.toLowerCase())
+      )
     );
   }, [search]);
 
@@ -61,16 +44,18 @@ export default function IconBuilder() {
     setIsDarkTheme(resolvedTheme === "dark");
   }, [resolvedTheme]);
 
-  const toggleIcon = (icon: string) => {
+  const toggleIcon = (icon: IconItem) => {
     setSelectedIcons((prev) =>
-      prev.includes(icon) ? prev.filter((i) => i !== icon) : [...prev, icon]
+      prev.some((i) => i.id === icon.id)
+        ? prev.filter((i) => i.id !== icon.id)
+        : [...prev, icon]
     );
   };
 
   const generateUrl = (): string => {
     const baseUrl = "https://skillicons.dev/icons";
-    const iconParams = selectedIcons.join(",");
-    const themeParam = isDarkTheme ? "" : "&theme=light";
+    const iconParams = selectedIcons.map(icon => icon.id).join(",");
+    const themeParam = iconTheme ? "" : "&theme=light";
     const centerParam = isCentered ? "&center=true" : "";
     const perlineParam = iconsPerLine !== 15 ? `&perline=${iconsPerLine}` : "";
 
@@ -83,8 +68,7 @@ export default function IconBuilder() {
 
   const generateMarkdown = () => {
     const url = generateUrl();
-    const iconParams = selectedIcons;
-    if (iconParams.length === 0) {
+    if (url === "No icons selected") {
       return "No icons selected";
     } else {
       return `[![My Skills](${url})](https://skill-icons-builder.vercel.app/)\n\nMy Skills`;
@@ -93,8 +77,7 @@ export default function IconBuilder() {
 
   const generateHtmlCentered = () => {
     const url = generateUrl();
-    const iconParams = selectedIcons;
-    if (iconParams.length === 0) {
+    if (url === "No icons selected") {
       return "No icons selected";
     }
     return `<p align="center">
@@ -103,8 +86,6 @@ export default function IconBuilder() {
     </a>
   </p>`;
   };
-
-  const [isCopied, setIsCopied] = useState(false);
 
   const handleCopy = (format: "markdown" | "html" | "url") => {
     const textToCopy = format === "markdown" ? generateMarkdown() : format === "html" ? generateHtmlCentered() : generateUrl() || "";
@@ -117,10 +98,15 @@ export default function IconBuilder() {
   return (
     <div className="container mx-auto p-4 max-w-7xl mt-10">
       <div className="flex flex-col lg:flex-row gap-6">
-        <div className="w-full lg:w-1/3 lg:sticky lg:top-4 lg:self-start">
-          <h1 className="text-3xl font-bold mb-6 text-center" id="title">
-            Skill Icon Builder
-          </h1>
+        <div className="w-full lg:w-1/3 lg:sticky lg:top-10 lg:self-start">
+          <div className="mb-6 text-center" id="title">
+            <Image
+              src={`./Skill-Icons-Builder-Logo-${isDarkTheme ? 'light' : 'dark'}.svg`}
+              alt="Skill Icons Builder Logo"
+              width={500}
+              height={300}
+            />
+          </div>
           <div className="mb-6 flex gap-2">
             <Input
               type="text"
@@ -133,23 +119,45 @@ export default function IconBuilder() {
           </div>
           <Card className="p-4">
             <h2 className="text-xl font-semibold mb-4">Preview</h2>
-            <div className={`flex flex-wrap gap-2 min-h-[48px] ${isCentered && generateUrl() !== "No icons selected" ? "justify-center" : "justify-start"} ${generateUrl() === "No icons selected" ? "gray-500 dark:text-gray-400 text-sm justify-start" : ""}`} >
-              {generateUrl() !== "No icons selected" ? (
-                <img
-                  src={generateUrl() as string}
-                  alt="Selected Skills"
-                  className="max-w-full"
-                />
-              ) : (
-                "Select some icons to generate a preview"
-              )}
+            <div className={`w-full flex ${isCentered && selectedIcons.length > 0 ? "justify-center" : "justify-start"}`}>
+              <ReactSortable
+                list={selectedIcons}
+                setList={setSelectedIcons}
+                animation={200}
+                delay={2}
+                className={`grid gap-2  w-fit`}
+                style={{
+                  display: 'grid',
+                  gridTemplateColumns: `repeat(${Math.min(selectedIcons.length, iconsPerLine)}, minmax(0,1fr)) `,
+                  justifyContent: isCentered ? "center" : "start",
+                  alignItems: 'center',
+                  placeItems: isCentered ? 'center' : 'start',
+                }}
+              >
+                {selectedIcons.length > 0 ? (
+                  selectedIcons.map((icon) => (
+                    <div key={icon.id} className="cursor-move h-fit w-fit">
+                      <img
+                        src={`https://skillicons.dev/icons?i=${icon.id}&theme=${iconTheme ? 'dark' : 'light'}`}
+                        alt={icon.name}
+                      />
+                    </div>
+                  ))
+                ) : (
+                  <span className="text-gray-500 dark:text-gray-400 text-sm mb-[28px]">
+                    Select some icons to generate a preview
+                  </span>
+                )}
+              </ReactSortable>
             </div>
+
+
           </Card>
           <div className="flex flex-wrap items-center gap-4 mb-6 mt-10">
             <div className="flex items-center gap-2">
               <Switch
-                checked={isDarkTheme}
-                onCheckedChange={setIsDarkTheme}
+                checked={iconTheme}
+                onCheckedChange={setIconTheme}
                 id="theme-toggle"
               />
               <label htmlFor="theme-toggle">Dark Theme</label>
@@ -169,7 +177,7 @@ export default function IconBuilder() {
               <Slider
                 id="icons-per-line"
                 min={1}
-                max={50}
+                max={20}
                 step={1}
                 value={[iconsPerLine]}
                 onValueChange={([value]) => setIconsPerLine(value)}
@@ -230,15 +238,18 @@ export default function IconBuilder() {
         <div className="w-full lg:w-2/3">
           <div className="grid grid-cols-3 sm:grid-cols-4 md:grid-cols-6 gap-4 mb-6">
             {filteredIcons.map((icon) => (
-              <Card key={icon} className={`p-4 cursor-pointer transition-colors ${selectedIcons.includes(icon) ? "bg-primary text-primary-foreground" : ""}`} onClick={() => toggleIcon(icon)} >
+              <Card
+                key={icon.id}
+                className={`p-4 cursor-pointer transition-colors ${selectedIcons.some(i => i.id === icon.id) ? "bg-primary text-primary-foreground" : ""
+                  }`}
+                onClick={() => toggleIcon(icon)}
+              >
                 <img
-                  src={`https://skillicons.dev/icons?i=${icon}&theme=${selectedTheme}`}
-                  alt={icon}
+                  src={`https://skillicons.dev/icons?i=${icon.id}&theme=${selectedTheme}`}
+                  alt={icon.name}
                   className="w-12 h-12 mx-auto mb-2"
                 />
-                <p className="text-center text-sm">
-                  {icon}
-                </p>
+                <p className="text-center text-sm">{icon.name}</p>
               </Card>
             ))}
           </div>
